[migration] AddPhoto.tsx, WriteForm.tsx useState 타입 수정
